style(errors): rename error files to fix code smells
By convention, a file that exports only one class, function, or constant should be named for that
class, function or constant. Anything else may confuse maintainers.
