+++ errors/BadRequestError.js
@@ -1,4 +1,4 @@
-import BaseError from './base';
+import BaseError from './BaseError';
 
 export default class BadRequestError extends BaseError {
   constructor({
