+++ errors/BaseError.js
