+++ pages/api/ibge/uf/v1/[code].js
@@ -1,6 +1,6 @@
 import app from '@/app';
 import { getUfByCode } from '@/services/ibge';
-import NotFoundError from '@/errors/not-found';
+import NotFoundError from '@/errors/NotFoundError';
 
 const action = async (request, response) => {
   const { code } = request.query;
